+++ cdl/core/model/image.py
@@ -766,6 +766,7 @@ class ImageTypes(base.Choices):
 class NewImageParam(gds.DataSet):
     """New image dataset"""
 
+    hide_image_dtype = False
     hide_image_type = False
 
     title = gds.StringItem(_("Title"))
@@ -775,7 +776,9 @@ class NewImageParam(gds.DataSet):
     width = gds.IntItem(
         _("Width"), help=_("Image width (total number of columns)"), min=1
     )
-    dtype = gds.ChoiceItem(_("Data type"), ImageDatatypes.get_choices())
+    dtype = gds.ChoiceItem(_("Data type"), ImageDatatypes.get_choices()).set_prop(
+        "display", hide=gds.GetAttrProp("hide_image_dtype")
+    )
     itype = gds.ChoiceItem(_("Type"), ImageTypes.get_choices()).set_prop(
         "display", hide=gds.GetAttrProp("hide_image_type")
     )